refactor(grovedb-query): split aggregate_sum & aggregate_count_and_sum out of query.rs
Mirror the existing `aggregate_count.rs` layout — each aggregate
variant's Query helpers and validation live in their own sibling
module, keeping the `Query` core in `query.rs` focused on
general-purpose query plumbing.

**Moved out of `query.rs`:**

- `grovedb-query/src/aggregate_sum.rs` (new):
  `new_aggregate_sum_on_range`, `aggregate_sum_on_range`,
  `has_aggregate_sum_on_range_anywhere`, `validate_aggregate_sum_on_range`
  plus their tests (selector-walking + cross-aggregate orthogonality
  arm for the ACASOR inner rejection).

- `grovedb-query/src/aggregate_count_and_sum.rs` (new):
  `new_aggregate_count_and_sum_on_range`,
  `aggregate_count_and_sum_on_range`,
  `has_aggregate_count_and_sum_on_range_anywhere`,
  `validate_aggregate_count_and_sum_on_range` plus their tests
  (happy path, all rejection arms, subquery walking, dispatch error).

`query.rs` shrinks substantially as a result. Module declarations
registered in `grovedb-query/src/lib.rs` next to the existing
`mod aggregate_count`. The methods are still on `Query` via
`impl Query { ... }` blocks so no caller change is needed.

`cargo test -p grovedb-query --lib` — all 184 tests pass.
